+++ CHANGELOG.md
@@ -10,6 +10,14 @@
 
 - [plugin] added support for `DocumentSymbolProviderMetadata` [#10811](https://github.com/eclipse-theia/theia/pull/10811) - Contributed on behalf of STMicroelectronics
 
+## v1.24.0 - unreleased
+
+[1.24.0 Milestone](https://github.com/eclipse-theia/theia/milestone/32)
+
+<a name="breaking_changes_1.24.0">[Breaking Changes:](#breaking_changes_1.24.0)</a>
+
+ - [markers] `ProblemDecorator` reimplemented to reduce redundancy and align more closely with VSCode. `collectMarkers` now returns `Map<string, TreeDecoration.Data>`, `getOverlayIconColor` renamed to `getColor`, `getOverlayIcon` removed, `appendContainerMarkers` returns `void`.
+
 ## v1.23.0 - 2/24/2022
 
 [1.23.0 Milestone](https://github.com/eclipse-theia/theia/milestone/31)